refactor: 모듈 복수형으로 변경
