+++ src/apps/server/app.module.ts
@@ -12,7 +12,7 @@ import { AppController } from './app.controller';
 import { RedisModule } from '@liaoliaots/nestjs-redis';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { RedisConfigFactory } from '@modules/cache/redis/redis.factory';
-import { ExperienceModule } from './experience/experience.module';
+import { ExperienceModule } from './experiences/experience.module';
 import { ResumesModule } from './resumes/resumes.module';
 import { TestModule } from './test/test.module';
